requirements.txt
asgiref==3.7.2
cffi==1.16.0
crispy-bootstrap5==2023.10
cryptography==41.0.7
distlib==0.3.7
Django==4.2.7
django-admin-honeypot==1.1.0
django-crispy-forms==2.1
django-environ==0.11.2
django-jazzmin==2.6.0
filelock==3.13.1
platformdirs==3.11.0
pycparser==2.21
pymemcache==4.0.0
PyMySQL==1.1.0
python-decouple==3.8
python-dotenv==1.0.0
sqlparse==0.4.4
tzdata==2023.3
virtualenv==20.24.6
whitenoise==6.6.0
